# src/docs/swagger.yaml
definitions:
...
  models.BasicErrorRes:
    properties:
      error:
        type: string
    type: object
  models.BasicRes:
    properties:
      message:
        type: string
    type: object
...
  models.CreatePet:
    properties:
      age:
        type: integer
      behaviouralNotes:
        type: string
      breed:
        type: string
      certificate:
        type: string
      dietyPreferences:
        type: string
      gender:
        type: string
      healthInformation:
        type: string
      name:
        type: string
      type:
        type: string
      vaccinations:
        type: string
    type: object
...
  models.Pet:
    properties:
      age:
        type: integer
      behaviouralNotes:
        type: string
      breed:
        type: string
      certificate:
        type: string
      dietyPreferences:
        type: string
      gender:
        type: string
      healthInformation:
        type: string
      name:
        type: string
      ownerUsername:
        type: string
      type:
        type: string
      vaccinations:
        type: string
    type: object
...
  models.SearchFilter:
    properties:
      descending:
        type: boolean
      end_price_range:
        type: number
      end_time:
        type: string
      location:
        type: string
      max_rating:
        type: number
      min_rating:
        type: number
      page_number:
        type: integer
      page_size:
        type: integer
      q:
        type: string
      sort_by:
        description: price, rating, name
        type: string
      start_price_range:
        type: number
      start_time:
        type: string
    type: object
  models.SearchHistory:
    properties:
      search_date:
        type: string
      search_filters:
        $ref: '#/definitions/models.SearchFilter'
    type: object
...
  models.User:
    properties:
      address:
        type: string
      dateOfBirth:
        type: string
      defaultAccountNumber:
        type: string
      defaultBank:
        type: string
      email:
        type: string
      fullName:
        type: string
      id:
        type: string
      individualID:
        type: string
      password:
        type: string
      pets:
        items:
          $ref: '#/definitions/models.Pet'
        type: array
      phoneNumber:
        type: string
      profilePicture:
        items:
          type: integer
        type: array
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /user:
    get:
      consumes:
      - application/json
      description: Retrieve all user
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of users per page
        in: query
        name: per
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      summary: Get all user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update user information (authentication required)
      parameters:
      - description: User object that needs to be updated
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Update user information
      tags:
      - User
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve user information by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      summary: Get user by ID
      tags:
      - User
  /user/delete-bank-account:
    delete:
      consumes:
      - application/json
      description: Delete the bank account associated with the current user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Delete bank account
      tags:
      - User
  /user/get-search-history:
    get:
      description: Gets the search history of the current user
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              history:
                items:
                  $ref: '#/definitions/models.SearchHistory'
                type: array
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Get search history
      tags:
      - User
  /user/pets:
    post:
      consumes:
      - application/json
      description: Add a new pet for the user (authentication required)
      operationId: AddUserPetHandler
      parameters:
      - description: Pet object to be added
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/models.CreatePet'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Add a new pet for the user
      tags:
      - Pets
  /user/pets/{id}:
    get:
      consumes:
      - application/json
      description: Get user's pets by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              pets:
                items:
                  $ref: '#/definitions/models.Pet'
                type: array
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      summary: Get user's pets by user ID
      tags:
      - Pets
  /user/pets/{idx}:
    delete:
      consumes:
      - application/json
      description: Delete user's pet (authentication required)
      parameters:
      - description: Pet Index
        in: path
        name: idx
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Delete user's pet
      tags:
      - Pets
    put:
      consumes:
      - application/json
      description: Update user's pet information (authentication required)
      parameters:
      - description: Pet Index
        in: path
        name: idx
        required: true
        type: string
      - description: Pet object that needs to be updated
        in: body
        name: pet
        required: true
        schema:
          $ref: '#/definitions/models.Pet'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Update user's pet information
      tags:
      - Pets
  /user/pets/me:
    get:
      description: Get current user's pets (authentication required)
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            properties:
              pets:
                items:
                  $ref: '#/definitions/models.Pet'
                type: array
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Get current user's pets
      tags:
      - Pets
  /user/set-default-bank-account:
    post:
      consumes:
      - application/json
      description: Set default bank account for the current user
      parameters:
      - description: Default account number
        in: body
        name: default_account_number
        required: true
        schema:
          type: string
      - description: Default bank
        in: body
        name: default_bank
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Set default bank account
      tags:
      - User
  /user/uploadProfileImage:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a profile image for the current user
      parameters:
      - description: Profile image file to upload
        in: formData
        name: profileImage
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.BasicRes'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.BasicErrorRes'
      security:
      - ApiKeyAuth: []
      summary: Upload profile image
      tags:
      - User
swagger: "2.0"